requirements.txt
Django==1.6.5
MySQL-python==1.2.5
XlsxWriter==0.5.6
argparse==1.2.1
distribute==0.6.24
django-autofixture==0.9.1
django-bootstrap3-datetimepicker==2.2.3
django-crispy-forms==1.4.0
django-datatables-view==1.9
django-debug-toolbar==1.2.1
django-selectable==0.8.0
django-smart-selects==1.0.9
django-uuidfield==0.5.0
jdcal==1.0
jsonfield==0.9.22
numpy==1.8.1
openpyxl==2.0.5
pandas==0.14.1
python-dateutil==2.2
python-memcached==1.53
pytz==2014.4
rpy2==2.4.2
scipy==0.14.0
selenium==2.42.1
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
xlrd==0.9.3
xlwt==0.7.5
